% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{carriots.analytics.addColumn}
\alias{carriots.analytics.addColumn}
\title{Method to add a new column to the table}
\usage{
carriots.analytics.addColumn(colname = NULL, dataType = NULL, conn = NULL)
}
\description{
@param colName - Names of the new column
 @param type -  Type of the column( supported types con$dataTypes)
 @param conn - BAConnection object obtained from connect API
}
