% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{carriots.analytics.metadata}
\alias{carriots.analytics.metadata}
\title{set some meta data datasource connection}
\usage{
carriots.analytics.metadata(url, token, isParameterized)
}
\arguments{
\item{url}{- Envision server URL}

\item{token}{- Secret key for the datasource
obtained from the App}

\item{isParameterized}{- TRUE/FALSE based on the module}
}
\description{
set some meta data datasource connection
}
