% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{carriots.analytics.reload_dataSource}
\alias{carriots.analytics.reload_dataSource}
\title{Reload Datasource}
\usage{
carriots.analytics.reload_dataSource(baseUrl, token, apiKey)
}
\arguments{
\item{baseUrl}{- Envision server URL}

\item{token}{- Token for the datasource
obtained from the App}

\item{apiKey}{- ApiKey for the user/company}
}
\description{
Function to refresh the datasource in envision, which reflects
the latest update
}
