% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{carriots.analytics.updateFrame}
\alias{carriots.analytics.updateFrame}
\title{Method to update the data in the table}
\usage{
carriots.analytics.updateFrame(dataframe = NULL, colname = NULL,
  conn = NULL)
}
\description{
Update the table with the values provided in the data frame. Below is the
structure of the data frame for WhereClause
}
\details{
@param dataframe - DataFrame for where Clause
 @param colname -  colname in dataframe which is to be added in to table
 @param conn - BAConnection object obtained from connect API
}
