% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{getDataSourceMetaData}
\alias{getDataSourceMetaData}
\title{Get DataSource Meta data}
\usage{
getDataSourceMetaData(baseUrl, secretKey)
}
\arguments{
\item{baseUrl}{- Envision server URL}

\item{secretKey}{- Secret key for the datasource
obtained from the App}
}
\description{
This is a function to obtain the meta data of a
datasource from the envision server by providing the secret
key of the datasource
}
