% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{reload_dataSource}
\alias{reload_dataSource}
\title{Reload Datasource}
\usage{
reload_dataSource(baseUrl, secretKey)
}
\arguments{
\item{baseUrl}{- Envision server URL}

\item{secretKey}{- Secret key for the datasource
obtained from the App}
}
\description{
Function to refresh the datasource in envision, which reflects
the latest update
}
