% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{carriots.analytics.connect}
\alias{carriots.analytics.connect}
\title{Get datasource connection}
\usage{
carriots.analytics.connect()
}
\arguments{
\item{baseUrl}{- Envision server URL}

\item{secretKey}{- Secret key for the datasource
obtained from the App}
}
\description{
This is a function to obtain connection to a
datasource from the envision server by providing the token
obtained from the envision App for the specific datasource
}
