% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{carriots.analytics.load}
\alias{carriots.analytics.load}
\title{Method to load the table data}
\usage{
carriots.analytics.load(columns = columns, conn = NULL)
}
\arguments{
\item{columns}{- vector of selected column names}

\item{conn}{- BAConnection object obtained from connect API}
}
\description{
User can pass the selected column names as the data frame returned by this
method will have only the selected columns
}
