% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envision_connect.R
\name{carriots.analytics.update}
\alias{carriots.analytics.update}
\title{Method to update the data in the table}
\usage{
carriots.analytics.update(where = NULL, set = NULL, conn = NULL)
}
\description{
Update the table with the values provided in the data frame. Below is the
structure of the data frame for WhereClause
}
\details{
@param where - DataFrame for where Clause
 @param set -  DataFrame for Set Clause
 @param conn - BAConnection object obtained from connect API
}
